% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{corr2graph}
\alias{corr2graph}
\title{Correlation matrix to graph}
\usage{
corr2graph(R, n, alpha = 0.05, type = "marg", ...)
}
\arguments{
\item{R}{Correlation matrix.}

\item{alpha}{Significance level used to compute the correlation threshold.
By default, alpha = 0.05.}

\item{...}{arguments to be passed to or from other methods.}

\item{n.}{Sample size (i.e., the number of subjects).}

\item{type.}{Correlation type: "marg" for marginal correlation, and 
"cond" for conditional correlation.}
}
\value{
An igraph object.
}
\description{
Convert a correlation matrix to an igraph object.
}
