% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semLatent.R
\name{cplot}
\alias{cplot}
\title{Network community plotting utility}
\usage{
cplot(g, membership, l = layout.auto, global = FALSE)
}
\arguments{
\item{g}{An igraph object.}

\item{membership}{a vector of node memberships.}

\item{l}{igraph layout option.}

\item{global}{Logical value. If TRUE, the plot of the input graph 
(coloured by cluster membership) will be generated alongwith independent 
module plots. If the input graph is very large, plotting could be 
computationally intensive (by default, global = FALSE).}
}
\description{
Merge and plot network communities of a graph as single 
nodes.
}
\examples{
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
membership <- clusterGraph(graph, type = "tahc", size = 10)
G <- cplot(graph, membership, global = TRUE)

}
