% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{dagitty2graph}
\alias{dagitty2graph}
\title{Graph conversion from dagitty to igraph}
\usage{
dagitty2graph(dagi, verbose = FALSE, ...)
}
\arguments{
\item{dagi}{A graph as a dagitty object ("dag" or "pdag").}

\item{verbose}{A logical value. If TRUE, the output graph is shown
(for \code{graph2dagitty} only). This argument is FALSE by default.}

\item{...}{Currently ignored.}
}
\value{
An igraph object.
}
\description{
Convert a dagitty object to a igraph object.
}
\examples{

# Conversion from igraph to dagitty  (and viceversa)
dagi <- graph2dagitty(sachs$graph, verbose = TRUE)
graph <- dagitty2graph(dagi, verbose = TRUE)

}
\author{
Mario Grassi \email{mario.grassi@unipv.it}
}
