% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semWeights.R
\name{edgeweight.r2z}
\alias{edgeweight.r2z}
\title{Graph weighting via Fisher's r-to-z transformation of correlation 
coefficient}
\usage{
edgeweight.r2z(graph, data, group, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{Binary vector. This vector must be as long as the number 
of subjects. Each vector element must be 1 for cases and 0 for control 
subjects.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A weighted graph, as an igraph object, with two edge attributes: 
"zsign" and "pv". Attribute "zsign" may have values [-1, 0, +1], where 
0 is assigned if Pr(|z|) > 0.1, while the sign of the z-test (either 
positive or negative) is assigned when Pr(|z|) < 0.1. Attribute "pv" 
correspond to the z-test P-value.
}
\description{
Weight connections of a bidirected (covariance or correlation) 
graph using two-group Fisher's transform correlation test, using standard 
error correction according to Yuan et al. 2013. Edge weights correspond 
to the sign and P-value of the correlation group difference test.
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
# Return graph properties, take the largest component, and convert 
# grapNEL to igraph
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
# Transpose data matrix: 32 subjectx (rows) x 19726 genes (columns)
data <- t(FTLDu_GSE13162)
G <- edgeweight.r2z(graph = graph, data = data, group = group)
E(G)$pv
E(G)$zsign

}
\references{
Yuan Z, Liu H, Zhang X, Li F, Zhao J, Zhang F, Xue F (2013). 
From Interaction to Co-Association - A Fisher r-To-z Transformation-Based 
Simple Statistic for Real World Genome-Wide Association Study. 
PLoS One, 8(7): e70774. https://doi.org/10.1371/journal.pone.0070774
}
\seealso{
\code{\link{edgeweight.sem}} for weighting directed graphs
}
