% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semWeights.R
\name{edgeweight.sem}
\alias{edgeweight.sem}
\title{Graph weighting via moderated mediation trivariate model}
\usage{
edgeweight.sem(graph, data, group, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{Binary vector. This vector must be as long as the number 
of subjects. Each vector element must be 1 for cases and 0 for control 
subjects.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A weighted graph, as an igraph object, with two edge attributes: 
zsign and pvalue. Attribute zsign may have values [-1, 0, +1], where 0 
is assigned if Pr(|z|) > 0.1, while the sign of the z-test (either 
positive or negative) is assigned when Pr(|z|) < 0.1. Attribute pvalue 
correspond to the z-test P-value.
}
\description{
Weight connections of a directed graph using a trivariate 
model in which the group variable has a direct effect on both source 
and target node, and their interaction. Edge weights correspond to 
the sign and P-value of the z-test (= estimate/standardError) of the 
mean direct group effects plus group moderation effect on the relationship 
between source and target.
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
# Return graph properties, take the largest component, and convert 
# grapNEL to igraph
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
# Transpose data matrix: 32 subjectx (rows) x 19726 genes (columns)
data <- t(FTLDu_GSE13162)
G <- edgeweight.sem(graph = graph, data = data, group = group)
E(G)$pv
E(G)$zsign

}
\references{
Grassi M & Palluzzi F (in preparation). SEMgraph: An R Package for 
Pathway and Network Analysis of Genomics Data with Structural 
Equation Models (SEM). Journal of Statistical Software (xxxx)

Palluzzi F, Ferrari R, Graziano F, Novelli V, Rossi G, Galimberti D, 
Rainero I, Benussi L, Nacmias B, Bruni AC, Cusi D, Salvi E, Borroni B, 
Grassi M. (2017). A novel network analysis approach reveals DNA damage, 
oxidative stress and calcium/cAMP homeostasis-associated biomarkers 
in frontotemporal dementia. PLoS ONE 12(10): e0185797. 
https://doi.org/10.1371/journal.pone.0185797
}
\seealso{
\code{\link{edgeweight.cfa}} and \code{\link{edgeweight.r2z}} 
for weighting bidirected (or undirected) graphs
}
