% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{gplot}
\alias{gplot}
\title{Graph plotting with renderGraph}
\usage{
gplot(x, psize = 80, main = "", ...)
}
\arguments{
\item{x}{An igraph object.}

\item{psize}{Node size (by default, psize = 80).}

\item{main}{plot main title.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Wrapper for \code{\link[Rgraphviz]{renderGraph}}.
}
