% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{graph2bap}
\alias{graph2bap}
\title{Convert directed graphs to bow-free acyclic path diagrams (BAPs)}
\usage{
graph2bap(graph, ...)
}
\arguments{
\item{graph}{A directed graph as an igraph object.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A BAP as an igraph object.
}
\description{
Convert cycle-containing directed graphs into BAPs.
When converting a graph to a SEM, it is desirable to have a recursive
(i.e., acyclic) model, that can be viewed as a directed acyclic graph
(DAG). BAPs are DAGs extension, allowing the presence of bidirected 
edges that can be interpreted as covariances (or latent variables
connecting pairs of nodes). Although cycles are permitted by SEM
theory, it is possible to convert them into BAPs. The graph2bap 
function enables topological sorting of the input graph until a cycle
is found. The first node of the cycle is defined by the hierarchical 
order, and thus the last edge of the cycle is determined (namely, 
the terminal edge). The terminal directed edge of each cycle is changed 
into a bidirected one, converting the input directed graph to a BAP.
}
