% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{graph2dagitty}
\alias{graph2dagitty}
\title{Graph to dagitty DAG format}
\usage{
graph2dagitty(graph, ...)
}
\arguments{
\item{graph}{igraph object.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
An dagitty graph object.
}
\description{
Convert an igraph object to a dagitty directed acyclic 
graph (DAG).
}
