% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{graph2lavaan}
\alias{graph2lavaan}
\title{Graph to path diagram}
\usage{
graph2lavaan(graph, nodes = V(graph)$name, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{nodes}{Subset of nodes to be included in the path diagram.
By default, all the graph nodes will be included inside the path diagram.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A path diagram in lavaan syntax.
}
\description{
Convert an igraph object to a path diagram, specified 
using lavaan syntax.
}
