% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{gsa.extract}
\alias{gsa.extract}
\title{Gene set analysis extraction utility}
\usage{
gsa.extract(
  gsa,
  reference = reactome.pathways,
  lcc = TRUE,
  min.size = 5,
  max.size = 300,
  min.drns = 2,
  act.alpha = 0.05,
  inh.alpha = 0.05,
  adj.alpha = 0.05,
  ...
)
}
\arguments{
\item{gsa}{Gene Set Analysis object generated by 
\code{\link[SEMgraph]{SEMgsa}}.}

\item{reference}{Reference list of biological pathways. Every element of the 
list must be an igraph object. SEMgraph provides both Reacrome and KEGG 
pathways list. By default, reactome.pathways is used.}

\item{lcc}{A logical value. If TRUE (default) only the largest connected 
component of the resulting network is considered.}

\item{min.size}{Minimum required pathway size (default = 5).}

\item{max.size}{Maximum allowed pathway size (default = 300).}

\item{act.alpha}{Pathway activation significance level (default = 0.05).}

\item{inh.alpha}{Pathway inhibition significance level (default = 0.05).}

\item{adj.alpha}{Global cutoff over the adjusted perturbation p-value 
(default = 0.05).}

\item{...}{Currently ignored.}

\item{min.degs}{Minimum required number of DRNs (default = 2).}
}
\value{
A list of two objects:
\enumerate{
\item "model", The output model as an igraph object;
\item "clusters", A list containing model clustering results.
It includes tree more objects:
\itemize{
\item "membership", A vector reporting cluster membership for each node;
\item "fit", hidden model fitting as a lavaan object;
\item "dataHM", data matrix containing cluster scores (hidden variables) 
alongside the original data variables.
}
}
}
\description{
Extract differentially regulated nodes (DRNs) and interaction 
network(s) from the \code{\link[SEMgraph]{SEMgsa}} output.
The interaction network is obtained by the union of all the perturbed 
pathways.
}
\examples{

\donttest{

# Nonparanormal(npn) transformation

library(huge)
als.npn <- huge.npn(alsData$exprs)

# Selection of FTD-ALS pathways from kegg.pathways.Rdata

 paths.name <- c("MAPK signaling pathway",
                 "Protein processing in endoplasmic reticulum",
                 "Endocytosis",
                 "Wnt signaling pathway",
                 "Neurotrophin signaling pathway",
                 "Amyotrophic lateral sclerosis")

j <- which(names(kegg.pathways) \%in\% paths.name)

GSA <- SEMgsa(kegg.pathways[j], als.npn, alsData$group,
               method = "bonferroni", alpha = 0.05,
               n_rep = 1000)

gsa.pert <- gsa.extract(GSA, kegg.pathways)

}

}
\references{
Grassi M, Palluzzi F, Tarantino B (2022). SEMgraph: An R Package for Causal Network
Analysis of High-Throughput Data with Structural Equation Models.
Bioinformatics, 38 (20), 4829–4830 <https://doi.org/10.1093/bioinformatics/btac567>
}
\seealso{
\code{\link[SEMgraph]{SEMgsa}}
}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
