% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\docType{data}
\name{kegg.pathways}
\alias{kegg.pathways}
\title{KEGG signaling pathways collection}
\format{"kegg.pathways" is a collection of 214 KEGG signaling pathways, saved as graphNEL objects.}
\source{
\url{https://www.kegg.jp/kegg}
}
\usage{
kegg.pathways
}
\description{
Collection of KEGG signaling pathways stored as igraph objects.
}
\examples{
# Number of nodes per pathway
kegg.nodes <- unlist(lapply(kegg.pathways, graph::numNodes))
# Number of edges per pathway
kegg.edges <- unlist(lapply(kegg.pathways, graph::numEdges))
# Gene list per pathway
kegg.genes <- unlist(lapply(kegg.pathways, graph::nodes))
quantile(kegg.nodes)
quantile(kegg.edges)
length(unique(kegg.genes))  # Number of unique genes within the dataset
# Loading breast cancer KEGG network
i <- which(names(kegg.pathways) == "hsa05224_Breast_cancer")
bc.graph <- kegg.pathways[[i]]
bc.graph
}
\references{
Kanehisa M1, Goto S (1999). KEGG: kyoto encyclopedia of genes and genomes. Nucleic Acid Research 28(1): 27-30. https://doi.org/10.1093/nar/27.1.29
}
