% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggInteractome.R
\docType{data}
\name{kegg}
\alias{kegg}
\title{KEGG interactome}
\format{"kegg" is an igraph network object of 5139 nodes and 68708 edges corresponding to the union of KEGG pathways (version 0.7.1).}
\source{
\url{https://www.kegg.jp/kegg}
}
\usage{
kegg
}
\description{
Interactome generated by merging KEGG pathways (version 0.7.1).
}
\examples{
kegg
# KEGG degrees of freedom
vcount(kegg)*(vcount(kegg) - 1)/2 - ecount(kegg)
}
\references{
Kanehisa M1, Goto S (1999). KEGG: kyoto encyclopedia of genes and genomes. Nucleic Acid Research 28(1): 27-30. https://doi.org/10.1093/nar/27.1.29
}
