% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{lavaan2graph}
\alias{lavaan2graph}
\title{Path diagram to graph}
\usage{
lavaan2graph(model, directed = TRUE, psi = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{model}{Path diagram using lavaan syntax.}

\item{directed}{Logical value. If TRUE (default), edge directions from 
the path diagram will be preserved. If FALSE, the resulting graph will 
be undirected.}

\item{psi}{Logical value. If TRUE (default) covariances will be converted 
into bidirected graph edges. If FALSE, covariances will be excluded from 
the output graph.}

\item{verbose}{Logical value. If TRUE (default), a plot of the output 
graph will be generated. For large graphs, this could significantly 
increase computation time. If FALSE, graph plotting will be disabled.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
An igraph object.
}
\description{
Convert a path diagram, specified using lavaan syntax, 
to an igraph object.
}
\examples{

# Writing path diagram in lavaan syntax

model <- '
# path diagram
y1 ~ x1
y2 ~ x1
y3 ~ y1 + y2
y4 ~ x1 + y1 + y3
# covariances
y1 ~~ y2'

# Converting path diagram into graph
graph <- lavaan2graph(model)
graph

}
