% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{mergeNodes}
\alias{mergeNodes}
\title{Network nodes merging by a user-defined membership attribute}
\usage{
mergeNodes(graph, membership, ...)
}
\arguments{
\item{graph}{Network as an igraph object.}

\item{membership}{Cluster membership. A vector of cluster membership 
identifiers, where vector names correspond to network node names. 
Network clustering can be done using \code{\link[SEMgraph]{clusterGraph}}.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A network with merged nodes as an igraph object.
}
\description{
Merge groups of network nodes using a custom membership 
attribute (e.g., cluster membership).
}
\examples{
# Data loading
group <- c(rep(0, 17), rep(1, 15))
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)

# Extracting hidden modules with SEMfsr
fsr.uv <- SEMfsr(graph = graph, data = data, group = group, 
                 type = "ebc",
                 HM = "UV",
                 size = 15)

# Merging nodes into modules
cg <- mergeNodes(graph = graph, membership = fsr.uv$M)

}
\seealso{
\code{\link[SEMgraph]{clusterGraph}}
}
