% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semWeights.R
\name{seedweight}
\alias{seedweight}
\title{Seed extraction methods}
\usage{
seedweight(graph, data, group, h = 0.2, alpha = 0.05, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{Binary vector. This vector must be as long as the number 
of subjects. Each vector element must be 1 for cases and 0 for control 
subjects.}

\item{h}{A decimal numbed, corresponding to the prototype clustering 
distance measure (= 1 - abs(correlation)) cutoff. By default, h = 0.2 
(corresponding to an absolute correlation coefficient of 0.8).}

\item{alpha}{Significance level of the group effect over graph nodes. 
By default, alpha = 0.05.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
The input graph endowed with three new binary (1: seed, 
0: not-seed) vertex attibutes:
\enumerate{
\item "spvlm", P-value of the simple linear regression y ~ x (i.e., 
node ~ group)
\item "sprot", prototype seeds derived from 
\code{\link[protoclust]{protoclust}}
\item "smst", non-leaf nodes of the MST derived from the input network
}
}
\description{
Extract important nodes (i.e., seeds) from a network using 
three methods: group influence p-value, Minimum Spanning Tree (MST), 
and prototype node detection.
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
# Return graph properties, take the largest component, and convert 
# grapNEL to igraph
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
# Transpose data matrix: 32 subjectx (rows) x 19726 genes (columns)
data <- t(FTLDu_GSE13162)
G <- seedweight(graph = graph, data = data, group = group)
table(V(G)$spvlm)
table(V(G)$sprot)
table(V(G)$smst)

}
\seealso{
\code{\link[stats]{lm}} for direct effect group estimation 
on graph nodes; \code{\link[igraph]{mst}} using Prim's algorithm for 
weighted networs; \code{\link[protoclust]{protoclust}} for prototype 
nodes calculation
}
