% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{sem.mediation}
\alias{sem.mediation}
\title{SEM-based mediation analysis}
\usage{
sem.mediation(
  data,
  x1,
  y,
  model = NULL,
  group = NULL,
  fixed.x = TRUE,
  se = "robust.huber.white",
  std.x = FALSE,
  std.y = FALSE,
  std.m = FALSE,
  adjust.method = "BH",
  gene.symbols = FALSE,
  ...
)
}
\arguments{
\item{data}{A matrix whith rows corresponding to subjects, and
columns to potential mediatiors.}

\item{x1}{A numeric vector corresponding to the exposure factor.}

\item{y}{A numeric vector corresponding to the outcome variable.}

\item{model}{The user might specify a different model. To use the 
given mediation model, leave this argument to NULL (default).}

\item{group}{A vector to group subjects for multigroup analysis. 
By default, \code{group = NULL}.}

\item{fixed.x}{Logical value. If TRUE (default), the exposure factor 
(i.e., the exogenous variable) estimates are fixed to its sampling 
values. If FALSE, their estimates are considered as free parameters.}

\item{se}{Standard error estimation method. Use "standard" (default) 
for conventional standard errors, based on the inverse information 
matrix. Use "robust.huber.white" for pseudo ML Huber-White estimation.}

\item{std.x}{Logical value. If TRUE, the exposure factor is standardized 
(default = FALSE).}

\item{std.y}{Logical value. If TRUE, the outcome variable is standardized 
(default = FALSE).}

\item{std.m}{Logical value. If TRUE, the mediator is standardized 
(default = FALSE).}

\item{adjust.method}{One of the multiple testing correction methods 
from the \code{p.adjust} function. By default, the Benjamini-Hochberg 
("BH") adjuctment method is used.}

\item{gene.symbols}{Logical value. If variables are genes and data 
matrix column names are entrez gene IDs, this argument can be set to 
TRUE to obtain a new column with official gene symbols 
(default = FALSE).}

\item{...}{Currently ignored.}
}
\value{
A list of two objects:
\enumerate{
\item "model", The output model as an igraph object;
\item "clusters", A list containing model clustering results.
It includes tree more objects:
\itemize{
\item "membership", A vector reporting cluster membership for each node;
\item "fit", hidden model fitting as a lavaan object;
\item "dataHM", data matrix containing cluster scores (hidden variables) 
alongside the original data variables.
}
}
}
\description{
Perform mediation analysis through structural equation 
modeling (SEM). The SEM is represented by two equations: 
x2 ~ b*x1 and y ~ a*x1 + c*x2, where x1 is the exposure factor, 
y is the outcome variable, and x2 is the mediator.
This model is iteratively applied for each column in the input 
data matrix, given x1 and y. A new parameter IE := b*c is defined, 
corresponding to the mediation effect.
}
\examples{

\donttest{

library(huge)
library(org.Hs.eg.db)

als.npn <- huge.npn(alsData$exprs)

# Random sampling 10 variables (possible mediators)
j <- sample.int(ncol(als.npn), 10)

# Generate a random binary exposure vector
exposure <- rbinom(n = nrow(als.npn), size = 1, prob = 0.3)

# Generate a random gaussian outcome vector
outcome <- rnorm(n = nrow(als.npn))

mediators <- sem.mediation(data = als.npn[, j],
                           x1 = exposure,
                           y = outcome,
                           gene.symbols = TRUE)

}

}
\author{
Fernando Palluzzi \email{fernando.palluzzi@gmail.com}
}
