% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMace}
\alias{SEMace}
\title{Compute Average Causal Effect (ACE) for a given DAG}
\usage{
SEMace(graph, data, group = NULL, ...)
}
\arguments{
\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the 
number of subjects. Each vector element must be 1 for cases and 0 
for control subjects. If NULL (default), group influence will not be 
considered.}

\item{...}{arguments to be passed to or from other methods.}

\item{model}{A directed acyclic graph as an igraph object.}
}
\value{
A list of 5 objects:
\enumerate{
\item "theta", data.frame containing total effect estimates;
\item "gD", graph representation of significant total effects.
}
}
\description{
Compute total effects as ACEs of source variables X on 
target variables in a DAG (i.e., when paths connecting X to Y are 
directed and acyclic). The ACE will be estimated as the path coefficient 
of X (i.e., theta) in the linear equation Y ~ X + pa(X), where pa(X) 
is the set of parent variables of X (Pearl, 1998).
}
\examples{
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)
group <- c(rep(0, 17), rep(1, 15))
ace <- SEMace(graph, data, group)
ace$theta

}
\references{
Pearl J (1998). Graphs, Causality, and Structural Equation Models. 
Sociological Methods & Research. https://doi.org/10.1177/0049124198027002004
}
\seealso{
\code{\link[SEMgraph]{SEMfit}}
}
