% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMbap}
\alias{SEMbap}
\title{Bow-free covariance search}
\usage{
SEMbap(graph, data, gnet = NULL, d = 0, alpha = 0.05, verbose = FALSE, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix whith rows corresponding to subjects, and
columns to graph nodes (variables).}

\item{gnet}{Reference "global" network as an igraph object.}

\item{d}{An integer value indicating the maximum length of indirect
interactions between pairs of nodes. If d = 1, direct interactions 
between nodes will be searched in the reference interactome (if given). 
If d > 1, indirect interactions of length d or shorter (i.e., with at 
most d - 1 connectors) between bow-free nodes will be searched. 
Setting d = 0, is equivalent to gnet = NULL.}

\item{alpha}{Significance level for Bonferroni multiple testing correction 
used for the independence tests (by default, alpha = 0.05), as suggested 
by Shipley (2000).}

\item{verbose}{A logical value. If FALSE (default), the processed graphs
will not be plotted to screen, saving execution time.}

\item{...}{Currently ignored.}
}
\value{
A list of 3 igraph objects:
\itemize{
\item "bap", the output bow-free acyclic path diagram,
\item "guu", the bidirected graph of significant covariances,
\item "gLV", the directed graph of latent variables (LV) underlying
significant covariances (i.e., the canonical graph, where bidirected
X <-> Y edges are substituted by directed edges X <- LV -> Y).
}
}
\description{
Search for new bow-free interactions. SEMbap does an 
exhaustive missing interaction search, using Shipley's d-separation test 
(Shipley, 2000) to add a bidirected edge (i.e., a bow-free covariance) 
between pairs of disconnected nodes (X, Y) if there is a significant 
association between them, given a set Z of conditioning nodes. 
Here, the "bow-free" condition states that there cannot be both 
directed and bidirected connections between the same pair of nodes 
(Nowzohour, 2017).
}
\details{
Shipley's d-separation test (Shipley, 2000) evaluates if two 
variables (X, Y) in a DAG are conditionally independent for a given 
conditioning set Z, assuming that the underlying statistical model is 
recursive (i.e., acyclic) and can be represented as a graph in which 
nodes (i.e., variables) are connected through directed edges (i.e., 
causal effects), with no self-loops. The conditioning set Z is 
represented by the union of the parent sets of X and Y (i.e., the 
"basis set"). SEMbap algorithm makes an exhaustive search of all possible 
missing edges, corresponding to the number of degrees of freedom (df) 
of the DAG. A new bow-free covariance is added if there is a significant 
X-Y association, after Bonferroni multiple testing correction. In addition, 
if a reference network is given, new interactions will be retained 
only if present in the reference. 
SEMbap function may work on any kind of directed network, by converting 
mixed graphs into directed acyclic graphs. 
See \code{\link[SEMgraph]{graph2dag}} for conversion method details.
}
\examples{

# BAP estimation
G <- SEMbap(graph = sachs$graph, data = log(sachs$pkc))

# Model fitting
sem <- SEMrun(graph = G$bap, data = log(sachs$pkc), group = sachs$group)

# Graphs
gplot(G$bap)
plot(G$guu)
plot(G$gLV)

}
\references{
Shipley B (2000). A new inferential test for path models based on DAGs. 
Struct. Equ. Modeling, 7(2): 206-218. 
https://doi.org/10.1207/S15328007SEM0702_4

Brito C and Pearl J (2002). A New Identification Condition for 
Recursive Models With Correlated Errors. 
Structural Equation Modeling, 9(4): 459-474.

Nowzohour C, Maathuis MH, Evans RJ, Buhlmann P (2017). Distributional 
equivalence and structure learning for bow-free acyclic path diagrams. 
Electron. J. Stat., 11(2): 5342-5374.
https://doi.org/10.1214/17-EJS1372
}
