% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMdsep}
\alias{SEMdsep}
\title{Search and test for causal relationships in a graph model}
\usage{
SEMdsep(fit, psi = FALSE, k = 3, MI = FALSE, alpha = 0.05,
  method = "BH", ...)
}
\arguments{
\item{fit}{A SEM fitted model object of class \code{\link[lavaan]{lavaan}}.}

\item{psi}{Logical value. If FALSE (default), covariances will be excluded 
from the output graph. If TRUE, covariances will be converted into 
bidirected graph edges.}

\item{k}{Numeric. Maximum number of conditioning variables in the 
dependence set. By default, k is set to the degrees of freedom of the 
graphical model. By default, k is set to 3. Increasing k may significantly 
increase computational time.}

\item{MI}{Logical value. If TRUE, new causal relationships are filtered 
by significant modification indices. By default, MI = FALSE. Enabling MIs 
will sensibly increase computational time.}

\item{alpha}{Numeric. Significance level for edge inclusion.}

\item{method}{P-value correction method. It may take the p.adjust.methods 
from \code{\link[stats]{p.adjust}} function. By default, it is set to 
"BH" ("Benjamini-Hochberg").}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A data.frame of causal interactions between d-separated variables.
}
\description{
Test the presence of causal relationships in a graphical 
model. The set of conditionally-independent variables are computed 
according to either Shipley's (Shipley, ...) or Pearl's (Pearl, ...) 
d-separation test, for directed acyclic graphs (DAGs) and non-DAGs, 
respectively. SEM modification indices (MIs) can be used to assign 
directions (i.e., causality) to the significantly interacting variables 
found in the first step.
}
\examples{
...

}
\references{
REF1

REF2
}
