% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMfit}
\alias{SEMfit}
\title{Fit a network as a Structural Equation Model (SEM)}
\usage{
SEMfit(graph, data, group = NULL, B = NULL, perm = 0, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix whith rows corresponding to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the 
number of subjects. Each vector element must be 1 for cases and 0 
for control subjects. If NULL (default), group influence will not be 
considered.}

\item{B}{Node-node interaction fixed weight. If B is NULL (default), 
beta coefficients will be estimated by MLE. If B is numeric, it will 
be used as a scaling factor for the edge weights in the graph object 
(graph attribute E(graph)$weight). Since SEMgraph scales data before 
model fitting, we suggest a grid search for the optimal B value in 
the interval [0, 0.3]. As a rule of thumb, to our experience, 
B = 0.1 performs well on any network.}

\item{perm}{Number of permutations. By default, perm is set to 0 and 
conventional standard errors will be computed. If perm > 1, P-values 
will be computed from a moment-based chi-squared approximation 
derived from the empirical distribution of permuted data (Larson and 
Owen, 2015). To reduce computational time costs per permutation, we 
suggest perm = 500 (this will leave P-values precision almost 
unaltered). In perm = 1, no P-values are calculated.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list of 5 objects:
\enumerate{
\item "fit", SEM fitted object of class lavaan;
\item "gest", group effect estimates and P-values on subgraph nodes;
\item "model", SEM model as a string;
\item "graph", the induced subgraph of the input network mapped on 
data variables;
\item "dataXY", input data subset mapping graph nodes, plus group at 
the first column (if no group is specified, this column will take NA 
values).
}
}
\description{
SEMfit converts a network with directed (x -> y), 
undirected (x -- y), or mixed (directed: x -> y and bidirected: 
x <-> y) edges to a SEM, and fits it through the 
\code{\link[lavaan]{lavaan}} function, via Maximum Likelihood 
Estimation (estimator = "ML", default estimator in 
\code{\link[lavaan]{lavOptions}}). If a binary group 
(i.e., case/control) variable is present, a group effect is added to 
the model as an exogenous variable. SEMfit can handle loops, 
although double links are not allowed, such as: self-loops, direct 
feedbacks, and bow (i.e., double directed/bidirected) links. If a 
directed or mixed network is given as input, SEMfit outputs three main 
sets of parameter estimates: (i) group effect estimations on network 
nodes (gamma), and aggregted group effects: D = sum of group 
effects, adjusted by the residual variance; A = sum of the tagret 
nodes perturbation (i.e., group effect) accumulation from source 
nodes; and E = sum of the source nodes perturbation (i.e., group 
effect) emission towards target nodes; (ii) beta coefficients (network 
interaction effects), and (iii) residual (co)variances (psi).
When the group variable is present, beta and psi coefficients, common 
to both groups, will be estimated. If the input is an undirected 
network, only psi coefficients will be estimated. In the latter case, 
two sets of parameters will be estimated: (i) group effect estimations 
on network nodes (gamma), and aggregted group effects: D = sum of group 
effects, adjusted by residual variances and covariances; V = sum of 
group effects, adjusted by the residual variances; and C = sum of 
the between-node covariance perturbation (i.e., group effect); and 
(ii) variance and covariance common to both groups.
P-values for parameter sets are computed by conventional z-test
(= estimate/SE), through \code{\link[lavaan]{lavaan}}. However, for 
very large networks, permutation test P-values can be computed 
(see \code{\link[flip]{flip}} for details).
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
# Return graph properties, take the largest component, and convert 
# grapNEL to igraph
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
# Transpose data matrix: 32 subjectx (rows) x 19726 genes (columns)
data <- t(FTLDu_GSE13162)

# Conventional standard errors computation
start <- Sys.time()
fit1 <- SEMfit(graph, data, group, B = NULL, perm = 0)
end <- Sys.time()
time1 <- end - start

# Moment-based chi-squared approximation (Larson and Owen, 2015)
start <- Sys.time()
fit2 <- SEMfit(graph, data, group, B = NULL, perm = 10000)
end <- Sys.time()
time2 <- end - start

# Execution time comparison
time1
time2

# if perm = 0 (i.e., canonical SE computation - no permutations):
# SEM goodness of fit indices
fitMeasures(fit1$fit, c("rmsea", "rmsea.pvalue", "srmr"))
R2 <- inspect(fit1$fit, "rsquare"); R2; mean(R2)
# SEM parameter estimates
summary(fit1$fit)
parameterEstimates(fit1$fit)[1:10,]
# Group effect estimates and aggregated group effects (i.e., D, A, E)
fit1$gest
pvalue <- fit1$gest$pvalue[-c(1:3)]
# Multiple testing correction
length(which(p.adjust(pvalue, method = "BH") < 0.05))

# if perm > 1 (i.e., moment-based chi-squared approximation):
# SEM goodness of fit indices
fitMeasures(fit2$fit, c("rmsea", "rmsea.pvalue", "srmr"))
R2 <- inspect(fit2$fit, "rsquare"); R2; mean(R2)
# SEM parameter estimates
summary(fit2$fit)
parameterEstimates(fit2$fit)[1:10,]
# Group effect estimates and aggregated group effects (i.e., D, A, E)
fit2$gest
pvalue <- fit2$gest@res$pchisq[-c(1:3)]    # chi-squared P-value
# Multiple testing correction
length(which(p.adjust(pvalue, method = "BH") < 0.05))
# Plotting flip permutation space
par(mfrow = c(2, 2))
flip::plot(fit2$gest[-c(1:3)])
flip::plot(fit2$gest[1, 2])
flip::plot(fit2$gest[1, 3])
flip::plot(fit2$gest[2, 3])

}
\references{
Yves Rosseel (2012). lavaan: An R Package for Structural Equation 
Modeling. Journal of Statistical Software, 48(2): 1-36.
URL http://www.jstatsoft.org/v48/i02/

Pepe D, Grassi M (2014). Investigating perturbed pathway modules 
from gene expression data via Structural Equation Models. BMC 
Bioinformatics, 15: 132.
URL https://doi.org/10.1186/1471-2105-15-132

Larson JL and Owen AB (2015). Moment based gene set tests. BMC 
Bioinformatics, 16: 132. https://doi.org/10.1186/s12859-015-0571-7
}
\seealso{
See \code{\link[lavaan]{lavaan}} for SEM fitting, and 
\code{\link[flip]{flip}} for permutation.
}
