% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMfit2}
\alias{SEMfit2}
\title{Fit a network as a two-group SEM and test group influence on 
network edges}
\usage{
SEMfit2(graph, data, group, fit = list(), ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the 
number of subjects. Each vector element must be 1 for cases and 0 
for control subjects.}

\item{fit}{Optional argument. In place of the arguments graph, data 
and group, a fitted model object returned by \code{\link{SEMfit}} or 
\code{\link{SEMbap}} functions, can be specified.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list of two objects: a fitted model object of class lavaan 
("fit"), and the estimate group differences ("dest") of the beta 
coefficients, covariances, or both if the input model is a directed, 
undirected, or mixed graph, respectively.
}
\description{
SEMfit2 converts an input network to a SEM and fits it
within each group (i.e., two-group SEM). SEMfit2 outputs two sets of
parameter estimates for each group: beta coefficients (network
interaction effects) and residual (co)variances if the input is a 
directed or mixed graph (see \code{\link[SEMgraph]{SEMfit}} for 
details), or variances and covariances if the input is an undirected 
graph.

Instead of modeling group effects as an exogenous variable, SEMfit2 
estimates the differences of the beta and/or psi coefficients 
(network edges) between groups.
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)

# Fitting from network data
fit2 <- SEMfit2(graph, data, group)
summary(fit2$fit)
dest <- fit2$dest
# Multiple testing correction
length(which(p.adjust(dest[, 7], method = "BH") < 0.05))

# Fitting from SEMfit output
fit1 <- SEMfit(graph, data, group, B = NULL, perm = 10000)
fit2 <- SEMfit2(fit = fit1)
summary(fit1$fit)
dest <- fit1$dest
# Multiple testing correction
length(which(p.adjust(dest[, 7], method = "BH") < 0.05))

}
\references{
Yves Rosseel (2012). lavaan: An R Package for Structural Equation 
Modeling. Journal of Statistical Software, 48(2), 1-36. 
URL http://www.jstatsoft.org/v48/i02/

Pepe D, Grassi M (2014). Investigating perturbed pathway modules 
from gene expression data via Structural Equation Models. 
BMC Bioinformatics, 15: 132. 
URL https://doi.org/10.1186/1471-2105-15-132
}
\seealso{
\code{\link[lavaan]{lavaan}}
}
