% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semLatent.R
\name{SEMfsr}
\alias{SEMfsr}
\title{Hidden module scoring and fitting via Factor Score Regression (FSR)}
\usage{
SEMfsr(graph, data, group, type, HM, size = 3, verbose = FALSE, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the number 
of subjects. Each vector element must be 1 for cases and 0 for control 
subjects.}

\item{type}{Graph clustering method, from \code{\link{igraph}} R package 
(see also \code{\link[SEMgraph]{clusterGraph}}).}

\item{HM}{Hidden model type. For each defined hidden module: 
(i) if HM = "LV", a latent variable (LV) will be defined as common 
unknown cause acting on cluster nodes; (ii) if HM = "CV", cluster nodes 
will be considered as regressors of a latent composite variable (CV); 
(iii) if HM = "UV", an unmeasured variable (UV) model will be generated 
for each latent module, where source nodes (i.e., in-degree = 0) act 
as common regressors influencing the other nodes via an unmeasured variable.}

\item{size}{Minimum number of nodes per hidden module. By default, a 
minimum number of 3 nodes is required.}

\item{verbose}{A logical value. If FALSE (default), the processed graphs 
will not be plotted to screen, saving execution time (they will be 
returned anyway).}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list of 3 objects:
\enumerate{
\item "fit", hidden module fitting as a lavaan object;
\item "M", hidden module nodes membership;
\code{\link[SEMgraph]{clusterGraph}} function;
\item "dataHM", hidden module data matrix.
}
}
\description{
Generate factor scores, principal component scores, or 
projection scores of latent, composite, and unmeasured variable modules, 
respectively, and fit them in a SEM with exogenous group effect.
}
\examples{
# Data loading
group <- c(rep(0, 17), rep(1, 15))
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)

# Module finding and Factor Score Regression
fsr.uv <- SEMfsr(graph = graph, data = data, group = group,
                 type = "ebc",
                 HM = "LV",
                 size = 15,
                 verbose = TRUE)

# Output summary
summary(fsr.uv$fit)
# Group membership
table(fsr.uv$M)
# Module scores
head(fsr.uv$dataHM)

# Hidden modules
gHC <- fsr.uv$gHC
gHC
gplot(x = gHC$g)

# Hidden modules reduction and fitting pipeline
fit <- SEMfit(graph, data, group, B = NULL, perm = 10000)
ggm <- SEMggm(fit = fit, gnet = kegg, d = 2,
              perm = 10000,
              alpha = 5E-05,
              verbose = FALSE)
ext <- SEMext(fit = ggm, gnet = kegg, data = data, B = NULL, d = 2,
              perm = 10000,
              alpha = 5E-05,
              verbose = FALSE)
cg <- mergeNodes(graph = ext$graph$Ug, membership = fsr.uv$M)
properties(cg)

# Adding hidden module data to the original data
data2 <- cbind(fsr.uv$dataHM[, -1], data)
head(data2)[, 1:10]
fit.hm <- SEMfit(graph = cg, data = data2, group = group, perm = 10000)
summary(fit.hm$fit)
summary(fit.hm$gest)

}
\references{
Hoshino, T., & Bentler, P.M. (2013). Bias in factor score regression 
and a simple solution. In de Leon, A.R., & Chough, K.C. (Eds.). 
Analysis of Mixed Data: Methods & Applications. 
New York: Chapman and Hall/CRC

Bai, J; Li, K (2012). Statistical analysis of factor models of high 
dimension. Ann. Statist. 40 (2012), no. 1, 436-465. 
doi:10.1214/11-AOS966.

Davies PT & Tso M K-S (1982). Procedures for Reduced-Rank Regression. 
Journal of the Royal Statistical Society. Series C (Applied Statistics), 
Vol. 31, No. 3, pp. 244-255
}
\seealso{
\code{\link[cate]{factor.analysis}}
}
