% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{SEMgsa}
\alias{SEMgsa}
\title{SEM-based gene set analysis}
\usage{
SEMgsa(g = list(), data, group, method = "BH", alpha = 0.05, n_rep = 1000, ...)
}
\arguments{
\item{g}{A list of pathways to be tested.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and
columns to graph nodes (variables).}

\item{group}{A binary vector. This vector must be as long as the number
of subjects. Each vector element must be 1 for cases and 0 for control
subjects.}

\item{method}{Multiple testing correction method. One of the values 
available in \code{\link[stats]{p.adjust}}. By default, method is set 
to "BH" (i.e., Benjamini-Hochberg correction).}

\item{alpha}{Gene set test significance level. Alpha is set to 0.05
by default.}

\item{n_rep}{Number of randomization replicates (default = 1000).}

\item{...}{Currently ignored.}
}
\value{
A data.frame reporting the following information for each 
pathway in the input list:
\itemize{
\item "N.nodes", pathway size (number of nodes);
\item "N.DRNs", number of differential expressed genes within the pathway, 
after multiple test correction with Benjamini-Hochberg method;
\item "pD", significance of the sum of group effects, adjusted by the 
residual variance;
\item "pA", significance of the sum of tagret nodes perturbation 
(i.e., group effect) accumulation from source nodes;
\item "pE", significance of the sum of source nodes perturbation 
(i.e., group effect) emission towards target nodes;
\item "pvalue", Fisher's combined P-value of pD, pA, and pE.
}
}
\description{
Gene Set Analysis (GSA) via self-contained test for group 
effect on signaling (directed) pathways as SEM, evaluating overall pathway 
perturbation, perturbation emission from source nodes, and perturbation 
accumulation on target nodes. Approximate randomization test P-values 
of specific node and aggregated group effects will be computed. 
For directed graphs, they include: the sum of group effects adjusted 
by residual variances (D), the sum of the tagret nodes perturbation 
(i.e., group effect) accumulation from source nodes (A), and the sum 
of the source nodes perturbation emission towards target nodes (E). 
For undirected graphs, the sum of group effects, adjusted 
by residual variances (D), will be estimated.
}
\examples{

library(SEMdata)
library(huge)

als.npn <- huge.npn(alsData$exprs)

pathway.names <- c("Amyotrophic lateral sclerosis (ALS)",
                   "Notch signaling pathway")
j <- which(names(kegg.pathways) \%in\% pathway.names)
pathways <- kegg.pathways[j]
GSA <- SEMgsa(pathways, als.npn, alsData$group, method = "BH", n_rep = 5000)
head(GSA$gsa)

\dontrun{

# GSA over the entire KEGG database

# Node count for each pathway
ngs <- unlist(lapply(1:length(kegg.pathways),
              function(x) vcount(kegg.pathways[[x]])))

# Remove pathways with less than 5 nodes and more than 500 nodes
Blacklist <- which(ngs < 5 | ngs > 500)
KEGG <- kegg.pathways[-Blacklist]

# DAG conversion
dag <- lapply(KEGG, function(x) graph2dag(x, data = als.npn))

# Gene Set Analysis
res <- SEMgsa(KEGG, als.npn, alsData$group, method = "BH", n_rep = 5000)
head(res$gsa)
dim(res$gsa)
head(res$DRN)
length(res$DRN)

}

}
\author{
Mario Grassi \email{mario.grassi@unipv.it}
}
