% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMpath}
\alias{SEMpath}
\title{Test (shortest) paths between two nodes of a network}
\usage{
SEMpath(graph, data, group, from, to, path, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the 
number of subjects. Each vector element must be 1 for cases and 0 
for control subjects. If NULL (default), group influence will not be 
considered.}

\item{from}{Starting node name.}

\item{to}{Ending node name.}

\item{path}{If path = "all", all the paths between the two nodes will 
be included in the fitted model. If path = "short", only shortest paths 
will be considered.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list of two objects: a fitted model object of class 
\code{\link[lavaan]{lavaan}} ("fit"), and the extracted subnetwork as 
an igraph object ("graph").
}
\description{
Find and fit all the (shortest) paths between two nodes 
of a network.
}
\examples{
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)
group <- c(rep(0, 17), rep(1, 15))
spt <- SEMpath(graph, data, group,
               from = "5155",
               to = "7082",
               path = "short")

}
