% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{Shipley.test}
\alias{Shipley.test}
\title{Missing edge testing implied by a graph}
\usage{
Shipley.test(graph, data, verbose = TRUE, ...)
}
\arguments{
\item{graph}{A directed graph as an igraph object.}

\item{data}{A data matrix with subjects as rows and variables as 
columns.}

\item{verbose}{If TRUE, Shipley's test results will be showed to 
screen (default = TRUE).}

\item{...}{Currently ignored.}
}
\value{
A list of two objects: (i) the list of all d-separation tests 
over missing edges in the input DAG or BAP, and (ii) the DAG or BAP 
used to perform the Shipley test.
}
\description{
Compute all the P-values of the d-separation tests 
implied by the missing edges of a given graph. The results of every 
test is then combined using the Fisher’s statistic in an overall test 
(Shipley's test) of the fitted model C = -2*sum(log(P-value(k))), where 
C is distributed as a chi-squared variate with df = 2k, as suggested 
by Shipley (2000).
}
\examples{

library(huge)
als.npn <- huge.npn(alsData$exprs)

sem <- SEMrun(alsData$graph, als.npn)
C.test0 <- Shipley.test(sem$graph, als.npn)

adjData <- diagonalizePsi(alsData$graph, als.npn, algo = "d-sep",
                          method = "BH",
                          alpha = 0.05)
C.test1 <- Shipley.test(sem$graph, adjData)

}
