% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringInteractome.R
\docType{data}
\name{string}
\alias{string}
\title{STRING interactome}
\format{"string" is an igraph network object of 9725 nodes and 170987 edges corresponding to the STRING interactome (version 10.5).}
\source{
\url{https://string-db.org}
}
\usage{
string
}
\description{
STRING interactome version 10.5.
}
\examples{
string
# STRING degrees of freedom
vcount(string)*(vcount(string) - 1)/2 - ecount(string)
}
\references{
Szklarczyk D, Gable AL, Lyon D, Junge A, Wyder S, Huerta-Cepas J, Simonovic M, Doncheva NT, Morris JH, Bork P, Jensen LJ, von Mering C (2019). STRING v11: protein-protein association networks with increased coverage, supporting functional discovery in genome-wide experimental datasets. Nucleic Acids Res., 47: D607-613. https://doi.org/10.1093/nar/gky1131
}
