% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{USPT}
\alias{USPT}
\title{Union of the Shortest Path Tree (USPT) algorithm}
\usage{
USPT(graph, seed, eweight, alpha = 1, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{seed}{A vector containing seed node identifiers, either user-defined 
or yielded by \code{\link[SEMgraph]{seedweight}}.}

\item{eweight}{Edge weight type derived from 
\code{\link[SEMgraph]{edgeweight.cfa}}, 
\code{\link[SEMgraph]{edgeweight.r2z}}, 
\code{\link[SEMgraph]{edgeweight.sem}}, 
or from user-defined distances. This option determines the 
weight-to-distance transform used to reduce the input graph. If set 
to "none" (default), edge weights will be set to 1. If eweight = "kegg", 
repressing interactions (-1) will be set to 1 (maximum distance), neutral 
interactions (0) will be set to 0.5, and activating interactions (+1) 
will be set to 0 (minimum distance). 
If eweight = "zsign", all significant interactions will be set to 0 
(minimum distance), while non-significant ones will be set to 1. 
If eweight = "pvalue", weights (p-values) will be transformed to the 
inverse of negative base-10 logarithm. If eweight = "custom", the 
algorithm will use the distance measure specified by the user as "weight" 
edge attribute.}

\item{alpha}{Significance level to assess shortest paths significance. 
By default, alpha is set to 1 (i.e., all the shortest paths will be 
retained).}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A list of two igraph objects, namely "Gs" and "sG". 
The former is the induced subnetwork of the shortest path nodes. 
The latter is the union of the shortest paths (i.e., the USPT).
}
\description{
Generate a subnetwork as the union of the most significant 
shortest paths from an input network. Shortest paths are chosen on 
the base of the perturbation (group difference) of their connections.
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)

# Generating edge weights
graph1 <- edgeweight.sem(graph = graph, data = data, group = group)

# Generating seed weights
graph2 <- seedweight(graph = graph1, data = data, group = group)
seed <- V(graph2)$name[V(graph2)$spvlm == 1]  # select seed nodes

# Using -log10(P-values) as edge weights
edgeweight <- 1/(-log10(E(graph2)$pv))
U <- USPT(graph = graph2, seed = seed, eweight = edgeweight)
graph3 <- U[[1]]    # induced subnetwork of the shortest path nodes
graph4 <- U[[2]]    # union of the shortest paths tree (USPT)
# Plotting output graph
V(graph3)$color <- ifelse(V(graph3)$name \%in\% seed, "yellow", "lightblue")
gplot(graph3)
V(graph4)$color <- ifelse(V(graph4)$name \%in\% seed, "yellow", "lightblue")
gplot(graph4)

}
\references{
Meier J, Tewarie P, Van Mieghem P (2015). The Union of Shortest Path 
Trees of Functional Brain Networks. Brain Connectivity 5(9); 575-581. 
https://doi.org/10.1089/brain.2014.0330
}
