% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{graph2lavaan}
\alias{graph2lavaan}
\title{Graph to lavaan model}
\usage{
graph2lavaan(graph, nodes = V(graph)$name, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{nodes}{Subset of nodes to be included in the model.
By default, all the input graph nodes will be included in the output model.}

\item{...}{Currently ignored.}
}
\value{
A model in lavaan syntax.
}
\description{
Convert an igraph object to a model, specified using 
lavaan syntax.
}
\examples{
model <- graph2lavaan(sachs$graph)
cat(model)

}
