% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{properties}
\alias{properties}
\title{Graph properties summary and graph correction}
\usage{
properties(g)
}
\arguments{
\item{g}{Input network as an igraph object.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Network largest component without self-loops.
}
\description{
Produces a summary of network properties and returns 
the largest network component without self-loops.
}
\examples{
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]

}
