% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMbap}
\alias{SEMbap}
\title{Bow-free graphical lasso data-driven and knowledge-based 
interaction search}
\usage{
SEMbap(graph, data, group = NULL, B = NULL, perm = 0,
  type = "dsep", alpha = 0.05, method = "BH", verbose = FALSE, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{data}{A matrix whith rows corresponding to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the 
number of subjects. Each vector element must be 1 for cases and 0 
for control subjects. If NULL (default), group influence will not be 
considered.}

\item{B}{Node-node interaction fixed weight. If B is NULL (default), 
beta coefficients will be estimated by MLE. If B is numeric, it will 
be used as a scaling factor for the edge weights in the graph object 
(graph attribute E(graph)$weight). Since SEMgraph scales data before 
model fitting, we suggest a grid search for the optimal B value in 
the interval [0, 0.3]. As a rule of thumb, to our experience, B = 0.1 
performs well on any network.}

\item{perm}{Number of permutations. By default, perm is set to 0 and 
conventional standard errors will be computed. If perm > 1, P-values 
will be computed from a moment-based chi-squared approximation derived 
from the empirical distribution of permuted data (Larson and Owen, 2015). 
To reduce computational time costs per permutation, we suggest perm = 500 
(this will leave P-values precision almost unaltered). In perm = 1, 
no P-values are calculated.}

\item{alpha}{Significance level used for GGM search. Alpha is set to 
0.05 by default.}

\item{verbose}{A logical value. If FALSE (default), the processed graphs 
will not be plotted to screen, saving execution time.}

\item{...}{arguments to be passed to or from other methods.}

\item{gnet}{External interaction network as an igraph object. Interaction 
data from this network will be used to integrate additional interaction 
information inside the model. Two preset databases are available: 
(i) kegg, for KEGG signaling pathways (directed), and (ii) string, for 
STRING protein interactions (undirected). Direct and indirect interactions 
will be added on the base of the selected interactome. Note that, if 
the input graph is undirected, an undirected reference interactome 
is expected.}

\item{d}{An integer value indicating the maximum length of indirect 
interactions between pairs of bow-free nodes. If d = 1, direct 
interactions between bow-free nodes will be searched in the reference 
interactome. If d > 1, indirect interactions of length d or shorter 
(i.e., with at most d - 1 connectors) between bow-free nodes will be 
searched.}
}
\value{
A list of 5 objects:
\enumerate{
\item "fit", SEM fitted object of class lavaan;
\item "gest", group effect estimates and P-values on subgraph nodes;
\item "model", SEM model as a string;
\item "graph", list of four graphs:
\itemize{
\item "ig", the induced subgraph mapped on input data,
\item "guu", the undirected GGM graph,
\item "gLV", the directed latent GGM graph,
\item "Ug", union of graphs "ig" and "guu";
}
\item "dataXY", input data subset mapping graph nodes, plus group at 
the first column (if no group is specified, this column will take NA 
values).
}
}
\description{
Fast estimation of sparse inverse covariance using Gaussian 
Graphical Models (GGMs), with de-sparsified graphical lasso method 
(D-S_GL; see \code{\link[SILGGM]{SILGGM}}). Significant bow-free covariaces 
will be selected. Only covariances connecting nodes without direct links 
in the input graph (i.e., bow-free covariances) and present in the 
reference interactome (i.e., biologically validated) are added to the 
input graph. Once new covariances has been added, the resulting mixed 
network is fitted (see \code{\link[SEMgraph]{SEMfit}} for further details).
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
# Return graph properties, take the largest component, and convert 
# grapNEL to igraph
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
# Transpose data matrix: 32 subjectx (rows) x 19726 genes (columns)
data <- t(FTLDu_GSE13162)
# Network degrees of freedom
ndf <- vcount(graph)*(vcount(graph) - 1)/2 - ecount(graph)
ggm <- SEMggm(graph = graph, data = data, gnet = kegg, group = group,
              d = 2,
              perm = 10000,
              alpha = 1/ndf,
              verbose = TRUE)

# Results summary
summary(ggm$gest)
pval <- ggm$gest@res$pchisq[-c(1:3)]
length(which(p.adjust(pval, method = "BH") < 0.1))

# Plot output graphs
par(mfrow = c(2, 2), mar = c(1, 1, 1, 1))
ig <- ggm$graph$ig; E(ig)$weight <- 1
guu <- ggm$graph$guu
gLV <- ggm$graph$gLV
Ug <- ggm$graph$Ug; E(Ug)$weight <- 1
gplot(ig, main = "Input network")
plot(guu, main = "Covariance network")
plot(gLV, main = "Latent Variable network")
gplot(Ug, main = "Input + Covariance network")

}
\references{
Zhang R, Ren Z, Chen W (2018). SILGGM: An extensive R package for 
efficient statistical inference in large-scale gene networks. 
PLoS Comput. Biol., 14(8): e1006369. 
https://doi.org/10.1371/journal.pcbi.1006369 

Friedman J, Hastie T, Tibshirani R (2008). Sparse inverse covariance 
estimation with the graphical lasso. Biostatistics, 9(3): 432-441. 
https://doi.org/10.1093/biostatistics/kxm045
}
\seealso{
\code{\link[lavaan]{lavaan}}, \code{\link{igraph}}, 
\code{\link[SILGGM]{SILGGM}}
}
