% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semUtils.R
\name{SEMdci}
\alias{SEMdci}
\title{SEM-based differential causal inference (DCI)}
\usage{
SEMdci(graph, data, group, type = "none", method = "BH", alpha = 0.05, ...)
}
\arguments{
\item{graph}{Input network as an igraph object.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and
columns to graph nodes (variables).}

\item{group}{A binary vector. This vector must be as long as the number
of subjects. Each vector element must be 1 for cases and 0 for control
subjects.}

\item{type}{Average Causal Effect (ACE) with two-group, "parents" (back-door)
adjustement set, and "direct" effects (\code{type = "ace"}), or a topological
clustering methods (default \code{type = "none"}). If \code{type = "tahc"},
network modules are generated using the tree agglomerative hierarchical
clustering method. Other non-tree clustering methods from igraph package
include: "wtc" (walktrap community structure with short random walks),
"ebc" (edge betweeness clustering), "fgc" (fast greedy method), "lbc"
(label propagation method), "lec" (leading eigenvector method), "loc"
(multi-level optimization), "opc" (optimal community structure), "sgc"
(spinglass statistical mechanics).}

\item{method}{Multiple testing correction method. One of the values
available in \code{\link[stats]{p.adjust}}. By default, method is set
to "BH" (i.e., FDR multiple test correction).}

\item{alpha}{Significance level (default = 0.05) for edge set selection.}

\item{...}{Currently ignored.}
}
\value{
An igraph object.
}
\description{
Creates a network with perturbed edges obtained from 
the output of \code{\link[SEMgraph]{SEMrun}} with two-group and CGGM
solver, comparable to the algorithm 2 in Belyaeva et al (2021), or of
\code{\link[SEMgraph]{SEMace}}, comparable to the procedure in
Jablonski et al (2022).
To increase the efficiency of computations for large graphs, users can
select to break the network structure into clusters, and select the
topological clustering method (see \code{\link[SEMgraph]{clusterGraph}}).
The function \code{\link[SEMgraph]{SEMrun}} is applied iteratively on
each cluster (with size min > 10 and max < 500) to obtain the graph
with the full list of perturbed edges.
}
\examples{

\dontrun{

# Nonparanormal(npn) transformation
library(huge)
als.npn <- huge.npn(alsData$exprs)

g <- kegg.pathways[["MAPK signaling pathway"]]
G <- properties(g)[[1]]; summary(G)

# Create ALS network with perturbed edges using edge betweeness clustering
gU<- SEMdci(G, als.npn, alsData$group, type="ebc", method="BH", alpha=0.2)
gcU<- properties(gU)

old.par <- par(no.readonly = TRUE)
par(mfrow=c(2,2), mar=rep(1,4))
gplot(gcU[[1]], l="fdp") # max component
gplot(gcU[[2]], l="fdp") # 2nd cpmponent
gplot(gcU[[3]], l="fdp") # 3rd component 
gplot(gcU[[4]], l="fdp") # 4th component
par(old.par)

}

}
\references{
Belyaeva A, Squires C, Uhler C (2021). DCI: learning causal differences
between gene regulatory networks. Bioinformatics, 37(18): 3067–3069.
<https://doi: 10.1093/bioinformatics/btab167>

Jablonski K, Pirkl M, Ćevid D, Bühlmann P, Beerenwinkel N (2022).
Identifying cancer pathway dysregulations using differential
causal effects. Bioinformatics, 38(6):1550–1559.
<https://doi.org/10.1093/bioinformatics/btab847>
}
\author{
Mario Grassi \email{mario.grassi@unipv.it}
}
