% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{SEMgsa}
\alias{SEMgsa}
\title{SEM-based gene set analysis}
\usage{
SEMgsa(g = list(), data, group, alpha, ...)
}
\arguments{
\item{g}{A list of pathways to be tested.}

\item{data}{A matrix or data.frame. Rows correspond to subjects, and 
columns to graph nodes.}

\item{group}{A binary vector. This vector must be as long as the number 
of subjects. Each vector element must be 1 for cases and 0 for control 
subjects.}

\item{alpha}{Gene set test significance level. Alpha is set to 0.05 
by default.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
A data.frame reporting group effect estimates over nodes
}
\description{
Test group effect on biological pathways as SEM, evaluating 
overall pathway perturbation, perturbation emission from source nodes, 
and perturbation accumulation on target nodes.
}
\examples{
group <- c(rep(0, 17), rep(1, 15))
graph <- properties(kegg.pathways$hsa04540_Gap_junction)[[1]]
data <- t(FTLDu_GSE13162)

ids <- which(names(kegg.pathways) \%in\% c("hsa04010_MAPK_signaling_pathway",
                                         "hsa04310_Wnt_signaling_pathway",
                                         "hsa04330_Notch_signaling_pathway",
                                         "hsa04722_Neurotrophin_signaling_pathway",
                                         "hsa04141_Protein_processing_in_endoplasmic_reticulum",
                                         "hsa04144_Endocytosis"))

gsa <- SEMgsa(kegg.pathways[ids], data, group, alpha = 0.05)
gsa

}
\references{
Grassi M, Pepe D, Palluzzi F (in preparation). Structural Equation 
Models (SEM)-based assessment of signalling pathway topology and 
perturbation. Journal ---- (xxxx)
}
\seealso{
\code{\link[SEMgraph]{SEMfit}}
}
