% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPaths.R
\name{summary.GGM}
\alias{summary.GGM}
\title{GGM model summary}
\usage{
\method{summary}{GGM}(object, ...)
}
\arguments{
\item{object}{A constrained GGM fitted model object.}

\item{...}{Currently ignored.}
}
\description{
Generate a summary for a constrained Gaussian Graphical 
Model (GGM) and show it to standard output.
}
\examples{
sem1 <- SEMrun(sachs$graph, log(sachs$pkc), sachs$group, algo = "cggm")
summary(sem1$fit)

}
\seealso{
\code{\link[SEMgraph]{SEMrun}}.
}
\author{
Mario Grassi \email{mario.grassi@unipv.it}
}
